kul: update

+++ adk_feed/uninstall.php
@@ -1,13 +1,9 @@
 <?php
 
 // If uninstall is not called from WordPress, exit
-if ( !defined( 'WP_UNINSTALL_PLUGIN' ) ) {
-	exit();
-}
-
 function adk_rm( $item ) {
 	if ( is_dir( $item ) ) {
-		foreach( scandir( $item ) as $i ) {
+		foreach( scandir( $item, SCANDIR_SORT_ASCENDING ) as $i ) {
 			if ( '.' !== $i[ 0 ] ) {
 				adk_rm( $item . $i );
 			}
@@ -20,8 +16,13 @@ function adk_rm( $item ) {
 	}
 }
 
-//delete_option( 'adk_notification' );
+if ( !defined( 'WP_UNINSTALL_PLUGIN' ) ) {
+	exit();
+}
+
 global $wpdb;
+
+delete_option( 'adk_feed_last_update' );
 
 $wpdb->query( "DROP TABLE {$wpdb->prefix}adk_feed_data" );
